Use keybaseExec in chatbot
