+++ lib/chat/bot.js
@@ -1,5 +1,4 @@
 // @flow
-import {runApiCommand} from '../chat-api.js'
 import keybaseExec from '../keybase-exec'
 import type {ApiCommandArg} from '../chat-api.js'
 import type {
@@ -11,6 +10,8 @@ import type {
   ChatOptionsRead,
 } from '../types.js'
 import BaseBot from '../base-bot'
+import {formatOptions} from '../utils'
+import {CHAT_API_VERSION} from '../constants'
 
 /**
  * A Keybase bot.
@@ -24,7 +25,7 @@ class ChatBot extends BaseBot {
    * @memberof Bot
    */
   list(options: ChatOptionsList): Promise<any> {
-    return this._safelyRunApiCommand({method: 'list', options})
+    return this._runApiCommand({method: 'list', options})
   }
 
   /**
@@ -33,7 +34,7 @@ class ChatBot extends BaseBot {
    * @memberof Bot
    */
   send(options: ChatOptionsSend): Promise<any> {
-    return this._safelyRunApiCommand({method: 'send', options})
+    return this._runApiCommand({method: 'send', options})
   }
 
   /**
@@ -47,7 +48,7 @@ class ChatBot extends BaseBot {
       peek: options.peek !== undefined ? options.peek : false,
       unreadOnly: options.unreadOnly !== undefined ? options.unreadOnly : false,
     }
-    return this._safelyRunApiCommand({method: 'read', options: optionsWithDefaults})
+    return this._runApiCommand({method: 'read', options: optionsWithDefaults})
   }
 
   /**
@@ -58,7 +59,7 @@ class ChatBot extends BaseBot {
    * @memberof Bot
    */
   delete(options: ChatOptionsDelete): Promise<any> {
-    return this._safelyRunApiCommand({method: 'delete', options})
+    return this._runApiCommand({method: 'delete', options})
   }
 
   /**
@@ -129,15 +130,21 @@ class ChatBot extends BaseBot {
     })
   }
 
-  // --------------------------------------------------------------------------
-  //  - make sure inited ok
-  //  - make sure user is still the same user since init
-  // --------------------------------------------------------------------------
-
-  _safelyRunApiCommand(arg: ApiCommandArg): Promise<any> {
-    return this._checkUserAndInit().then(() => {
-      return runApiCommand(arg)
-    })
+  async _runApiCommand(arg: ApiCommandArg): Promise<any> {
+    await this._checkUserAndInit()
+    const options = formatOptions(arg.options)
+    const input = {
+      method: arg.method,
+      params: {
+        version: CHAT_API_VERSION,
+        options,
+      },
+    }
+    const inputString = JSON.stringify(input)
+    const size = inputString.length
+    const output = await keybaseExec(['chat', 'api'], {stdinBuffer: Buffer.alloc(size, inputString, 'utf8')})
+    const res = output.result
+    return res
   }
 }
